+++ packages/cinterop/src/darwin/kotlin/io/realm/interop/RealmInterop.kt
@@ -399,7 +399,7 @@ actual object RealmInterop {
                 )
             )
             if (!found.value) {
-                throw RuntimeException("Class \"$name\" not found")
+                throw IllegalArgumentException("Class \"$name\" not found")
             }
             return ClassKey(classInfo.key.toLong())
         }