+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/MutableRealmImpl.kt
@@ -20,6 +20,7 @@ import io.realm.Cancellable
 import io.realm.MutableRealm
 import io.realm.RealmList
 import io.realm.RealmObject
+import io.realm.interop.RealmCoreException
 import io.realm.interop.RealmInterop
 import io.realm.isFrozen
 import io.realm.isManaged
@@ -59,7 +60,11 @@ internal class MutableRealmImpl : BaseRealmImpl, MutableRealm {
         super(configuration, RealmInterop.realm_open(configuration.nativeConfig, dispatcher))
 
     internal fun beginTransaction() {
-        RealmInterop.realm_begin_write(realmReference.dbPointer)
+        try {
+            RealmInterop.realm_begin_write(realmReference.dbPointer)
+        } catch (exception: RealmCoreException) {
+            throw genericRealmCoreExceptionHandler("Cannot begin the write transaction", exception)
+        }
     }
 
     internal fun commitTransaction() {
@@ -90,7 +95,11 @@ internal class MutableRealmImpl : BaseRealmImpl, MutableRealm {
     }
 
     override fun cancelWrite() {
-        RealmInterop.realm_rollback(realmReference.dbPointer)
+        try {
+            RealmInterop.realm_rollback(realmReference.dbPointer)
+        } catch (exception: RealmCoreException) {
+            throw genericRealmCoreExceptionHandler("Cannot cancel the write transaction", exception)
+        }
     }
 
     override fun <T : RealmObject> copyToRealm(instance: T): T {