+++ test/base/src/androidTest/kotlin/io/realm/shared/SampleTests.kt
@@ -123,10 +123,7 @@ class SampleTests {
     @Test
     fun query_parseErrorThrows() {
         val objects: RealmResults<Sample> = realm.objects(Sample::class)
-        // FIXME Need appropriate error for syntax errors. Avoid UnsupportedOperationExecption as
-        //  in realm-java ;)
-        //  https://github.com/realm/realm-kotlin/issues/70
-        assertFailsWith<RuntimeException> {
+        assertFailsWith<IllegalArgumentException> {
             objects.query("name == str")
         }
     }